server/requirements.txt
python-dotenv
fastapi
typing
apscheduler
fastapi_restful
python-multipart
httpx
uvicorn
langchain
composio_core
crewai
langchain_openai
composio_crewai
pydantic_settings
logging
sqlalchemy
asyncpg
greenlet
typing_inspect
aio_pika
google-genai
tiktoken
faiss-cpu